Refactor AutosizeInput component
