+++ src/components/input/index.ts
@@ -1,1 +1,2 @@
-export { default } from './input';
+export { default as AutosizeInput } from './autosize-input';
+export { default as TextareaInput } from './textarea-input';
