+++ src/pages/proposals/forms/delegate/delegate-form.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import classNames from 'classnames';
 import Button from '../../../../components/button';
-import Input from '../../../../components/input';
+import { AutosizeInput } from '../../../../components/input';
 import { ModalFooter, ModalHeader } from '../../../../components/modal';
 import { useChainData } from '../../../../chain-data';
 import { useApi3Pool } from '../../../../contracts';
@@ -68,9 +68,8 @@ const DelegateVotesForm = (props: Props) => {
       <ModalHeader>Delegate my votes to:</ModalHeader>
 
       <div className={globalStyles.textCenter}>
-        <Input
+        <AutosizeInput
           type="text"
-          autosize
           placeholder="Enter address or ENS name here"
           value={delegationAddress}
           onChange={(e) => {
